moved restaurant login with modal view to seperate file, working on saving selected restaurant item to db
